Conflicts:
	ezshop.db
	src/test/java/it/polito/ezshop/acceptanceTests/TestEZShop.java

+++ src/main/java/it/polito/ezshop/classes/EZShopDB.java
@@ -274,7 +274,10 @@ public class EZShopDB {
 	 */
 	public boolean addProductType(ProductType productType) {
 		String sql = "INSERT INTO productTypes(id, quantity, location, note, productDescription, barCode, pricePerUnit) VALUES(?,?,?,?,?,?,?)";
-
+		
+		if(productType == null)
+			return false;
+		
 		try (PreparedStatement pstmt = connection.prepareStatement(sql)) {
 			pstmt.setInt(1, productType.getId());
 			pstmt.setInt(2, productType.getQuantity());
@@ -314,6 +317,9 @@ public class EZShopDB {
 			String newNote) {
 		String sql = "UPDATE productTypes SET productDescription=?, barCode=?, pricePerUnit=?, note=? WHERE id=?";
 
+		if(id < 0)
+			return false;
+		
 		try (PreparedStatement pstmt = connection.prepareStatement(sql)) {
 			pstmt.setString(1, newDescription);
 			pstmt.setString(2, newCode);
@@ -358,8 +364,26 @@ public class EZShopDB {
 	}
 
 	public boolean deleteProductType(Integer id) {
-		String sql = "DELETE FROM productTypes WHERE id=?";
+		String sql = "SELECT COUNT(*) AS tot FROM productTypes WHERE id=?";
+		boolean exists = false;
+
 		try (PreparedStatement pstmt = connection.prepareStatement(sql)) {
+			pstmt.setInt(1, id);
+			ResultSet rs = pstmt.executeQuery();
+
+			if (rs.getInt("tot") > 0)
+				exists = true;
+
+		} catch (SQLException e) {
+			System.err.println(e.getMessage());
+			return false;
+		}
+		
+		if(!exists)
+			return false;
+		
+		String sql2 = "DELETE FROM productTypes WHERE id=?";
+		try (PreparedStatement pstmt = connection.prepareStatement(sql2)) {
 			pstmt.setInt(1, id);
 			pstmt.executeUpdate();
 		} catch (SQLException e) {
